# repr: speed up decimal parsing by 100%