**repr: speed up decimal parsing by 100%**

Due to a Rust bug, checked multiplication of i128s is two orders of
magnitude slower than it should be. See rust-lang/rust#53023.  Work
around the issue by using a u128 to perform arithmetic while parsing.

```
parse_decimal/-135412353251
                        time:   [3.1434 us 3.1449 us 3.1467 us]
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe

parse_decimal/1.030340E11
                        time:   [2.1604 us 2.1613 us 2.1624 us]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

parse_decimal/-135412353251
                        time:   [28.306 ns 28.316 ns 28.330 ns]
                        change: [-99.083% -99.066% -99.045%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  5 (5.00%) low mild
  2 (2.00%) high mild
  12 (12.00%) high severe

parse_decimal/1.030340E11
                        time:   [33.069 ns 33.110 ns 33.150 ns]
                        change: [-98.474% -98.472% -98.470%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
```

----

**pgrepr,repr: fix some decimal parsing/encoding bugs**

Noticed while comparing the PostgreSQL implementation to ours.

Fix #4925.

----

**repr: further optimize decimal parsing**

Avoid checked arithmetic where limits on precision already ensures that
overflow cannot occur. Also iterate over bytes rather than chars, which
is a bit faster.

Combined, this yields another 25% speedup, from ~35ns to ~28ns, on the
parse_decimal microbenchmark.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/materializeinc/materialize/5343)
<!-- Reviewable:end -->