@benesch 2021-01-16
Sounds good! If we want more confidence, I was thinking @mjibson could look at the fix in the pgrepr package and @sploiselle could look at the parsing tweaks so that `+` and `e` (lowercase) were handled. Or we can just merge it! I'm off hiking until late Mon, so folks should feel free to run with this.