@frankmcsherry 2021-01-16
I'll just merge it, unless there is a risk. I'd be delighted to check out the perf impact in TPCH queries, for example!