@frankmcsherry 2021-01-16
I don't understand the test error. `--rewrite-results` errors out, but manually changing the failing test to
```
query error exceeds maximum decimal precision
SELECT '-100000000000000000000000000000000000000'::decimal
----
100000000000000000000000000000000000000
```
works (i.e. removing the negative sign in the output that is not produced). I can push that change, but I'll not until someone can ack what is going on here. :D