@frankmcsherry 2021-01-16
I'm guessing it has to do with the fact that the error presents the positive literal in its error message, and something about how SLT works test the intended output against the error message (which does not have the negative number, as it errors with a still-positive decimal string).

Seems like the fix could be "change the test" or "change the error message to put a negative sign in place"?