@frankmcsherry 2021-01-16
Ok, more sniffing around implies that the problem is that `sqllogictest` itself errors trying to parse the large negative number. If you change that to anything else (e.g. "7") the test passes. I'm trying to read through the SLT code to see where the glitch is .. but .. yeah. :)