@frankmcsherry 2021-01-16
Yeah lots of weird things going on. When I test
```
query error exceeds maximum decimal precision
SELECT '100000000000000000000000000000000000000'::decimal
----
100000000000000000000000000000000000000

query I
SELECT '5'::decimal
----
5
```
this passes. When I test (by mistake)
```
query error exceeds maximum decimal precision
SELECT '100000000000000000000000000000000000000'::decimal
----
100000000000000000000000000000000000000

query error exceeds maximum decimal precision
SELECT '5'::decimal
----
5
```
this also passes (internally, SLT kicks up a parse error 
```
ParserError { message: "Expected end of statement, found number", pos: 63 }
```
but still returns success because:
```
            Err(e) if output.is_err() => {
                println!("{:?}", e);
                return Ok(Outcome::Success)
            },
```
When I test 
```
query error exceeds maximum decimal precision
SELECT '-100000000000000000000000000000000000000'::decimal
----
-100000000000000000000000000000000000000
```
it does *not* kick up a parser error and return successfully, it instead mints the following statement:
```
Select(SelectStatement { 
    query: Query { 
        ctes: [], 
        body: Select(Select { 
            distinct: None, 
            projection: [
                Expr { 
                    expr: Op { 
                        op: "-", 
                        expr1: Cast { 
                            expr: Value(String("-100000000000000000000000000000000000000")), 
                            data_type: Other { 
                                name: ObjectName([Ident("numeric")]), 
                                typ_mod: [] 
                            } 
                        }, 
                        expr2: Some(Value(Number("100000000000000000000000000000000000000"))) 
                    }, 
                    alias: None 
                }
            ], 
                from: [], selection: None, group_by: [], having: None, options: [] 
        }), 
        order_by: [], 
        limit: None, 
        offset: None 
    }, 
    as_of: None 
})
```
which just seems like a massive mis-parse of the test (it seems like it parsed the negative result as a subtraction, and then explodes because it can't parse the positive number).

No clue man.